% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdhFunctions.R
\name{extractSeriesCatalog}
\alias{extractSeriesCatalog}
\title{extractSeriesCatalog (gdh.db.in)}
\usage{
extractSeriesCatalog(sites, siteCode = NULL, featureType = NULL,
  variableCode = NULL, methodID = NULL, sourceID = NULL)
}
\arguments{
\item{sites}{data.frame obtenido mediante getSites}

\item{siteCode}{código de sitio (optativo)}

\item{featureType}{tipo de objeto espacial  ('point' o 'area' , optativo)}

\item{variableCode}{código de variable (optativo)}

\item{methodID}{ID de método (optativo)}

\item{sourceID}{ID de fuente (optativo)}
}
\description{
Esta función extrae las definiciones de series temporales de la propiedad
seriesCatalog (JSON string) del data.frame obtenido mediante getSites.
Permite filtrar por siteCode, featureType, variableCode, methodID y sourceID.
Devuelve un data.frame con las filas como definiciones de series y columnas como sus propiedades.
El mismo sirve como parámetro de entrada para la función getValues
}
\examples{
drv<-dbDriver("PostgreSQL")
con<-dbConnect(drv, user="sololectura",host='10.10.9.14',dbname='meteorology')
sites<-getSites(con,featureType='point',north=-20,south=-25,east=-55,west=-60)
extractSeriesCatalog(sites,variableCode=2,methodID=1)
}
\keyword{series}
\keyword{seriesCatalog}
