% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdhFunctions.R
\name{getSites}
\alias{getSites}
\title{getSites (gdh.db.in package)}
\usage{
getSites(con, siteCode = NULL, featureType = NULL, north = NULL,
  south = NULL, east = NULL, west = NULL)
}
\arguments{
\item{con}{Conexión a base de datos (creada con dbConnect)}

\item{siteCode}{Código del sitio (int, opcional)}

\item{featureType}{tipo de objeto espacial ('point' o 'area', opcional)}

\item{north}{coordenada norte del recuadro (opcional)}

\item{south}{coordenada sur del recuadro (opcional)}

\item{east}{coordenada este del recuadro (opcional)}

\item{west}{coordenada oeste del recuadro (opcional)}
}
\description{
Esta función descarga los sitios (features) de la base de datos hidrometeorológica, junto con las
definiciones de series temporales asociados a los mismos. Incluye puntos (featureType='point') y
areas (featureType='area'). Devuelve un data.frame cuyas filas corresponden a los sitios y las
columnas a las propiedades, el cual sirve como parámetro de entrada para la función extractSeriesCatalog
(la propiedad seriesCatalog es un JSON con la definición de las series temporales). Se puede obtener el listado
completo o filtrar por siteCode, featureType o recuadro espacial (north,south,east,west)
}
\examples{
drv<-dbDriver("PostgreSQL")
con<-dbConnect(drv, user="sololectura",host='10.10.9.14',dbname='meteorology')
getSites(con,featureType='point',north=-20,south=-25,east=-55,west=-60)
}
\keyword{getSite}
\keyword{sites}
