% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdhFunctions.R
\name{getValues}
\alias{getValues}
\title{getValues (gdh.db.in package)}
\usage{
getValues(con, seriesCatalog, startDate, endDate)
}
\arguments{
\item{con}{conexi칩n con la base de datos (creada mediante dbConnect)}

\item{seriesCatalog}{data.frame obtenido de la funci칩n extractSeriesCatalog, list de par치metros de una serie  o vector de valores de seriesCode}

\item{startDate}{fecha inicial}

\item{endDate}{fecha final}
}
\description{
Esta funci칩n descarga los pares de fecha y valor para las series definidas en
el data.frame seriesCatalog (o un vector de valores de seriesCode), de la base de datos definida en con, entre las fechas
startDate y endDate. Devuelve un data.frame con los registros en las filas y las
propiedades en las columnas
}
\examples{
drv<-dbDriver("PostgreSQL")
con<-dbConnect(drv, user="sololectura",host='10.10.9.14',dbname='meteorology')
sites<-getSites(con,featureType='point',north=-20,south=-25,east=-55,west=-60)
seriesCatalog<-extractSeriesCatalog(sites,variableCode=2,methodID=1)
getValues(con,seriesCatalog,'2010-01-01','2015-01-01')
}
\keyword{timeSeries}
\keyword{values}
