% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdhFunctions.R
\name{plotValues}
\alias{plotValues}
\title{plotValues (gdh.db.in package)}
\usage{
plotValues(values, seriesCatalog = NULL, output = NULL, types = c("l"),
  width = 800, height = 700, invert = c(FALSE), startDate = NULL,
  endDate = NULL)
}
\arguments{
\item{values}{data.frame obtenido mediante getValues}

\item{seriesCatalog}{data.frame obtenido mediante extractSeriesCatalog (opcional)}

\item{output}{Archivo donde se imprima el gráfico (opcional). Por defecto abre un display x11}

\item{types}{vector Tipos de gráfico para cada variable ('l': línea, 'p': puntos, 's' escalones), en el orden de aparición (opcional)}

\item{width}{Ancho del gráfico (opcional)}

\item{height}{Altura del gráfico (opcional)}

\item{invert}{vector de boolean Determina si debe invertirse el eje Y para cada variable (opcional)}

\item{startDate}{fecha inicial (opcional)}

\item{endDate}{fecha final (opcional)}
}
\description{
Esta función grafica las series temporales obtenidas mediante getValues.
Admite hasta dos variables (variableName) distintas, sin límite para el número
de series de cada variable.
}
\examples{
drv<-dbDriver("PostgreSQL")
con<-dbConnect(drv, user="sololectura",host='10.10.9.14',dbname='meteorology')
sites<-getSites(con,featureType='point',north=-20,south=-25,east=-55,west=-60)
seriesCatalog<-extractSeriesCatalog(sites,variableCode=2,methodID=1)
values<-getValues(con,seriesCatalog,'2010-01-01','2015-01-01')
plotValues(values,seriesCatalog,output='~/tmp/plotValues.png')
}
\keyword{plot}
\keyword{timeSeries}
\keyword{values}
